project form sidebar stays closed if user closes it (details below)
Sidebar component sends toggle only on manual user toggle, not when open-@input
changes. This allows knowing whether the sidebar toggled because of
user behavior.

Initialize project form sidebar with undefined, so that undefined stands
for "user has not manually toggled the sidebar", false for "user has
manually closed the sidebar" and true for "user has manually opened the
sidebar".

Project form keeps sidebar closed, if user has manually closed it.
Otherwise, follow the existing rules for when to autoclose the sidebar.