rm modal service hide method
Fixes a design problem with the modal service; it allows showing
multiple modals, but keeps only one reference to the latest modal.
If there were multiple modals, only the latest one could be controlled.

By having only a show method, the responsibility of hiding is left to
the client.  This forces the developer to keep components agnostic and
also safe from other components using the modal service.